Jmax/lg 9107 remove fsm document capture hybrid flow step (#8292)
* Ripped out CaptureCompleteStep

* Removed document capture step

* Removed CaptureDocController

* Removed VerifyDocumentStatusAction

* Removed CaptureDocFlow

* Removed doc_auth_hybrid_mobile_controllers_enabled feature flag

* Fix syntax error in UploadStep

* Remove VerifyDocumentStatusAction from DocAuthFlow

* lint

* Remove FSM document_capture path spec helpers

* Remove capture_doc routes

* Remove reference to @step_url in _document_capture.html.erb

This was used by the async code, which is unused.

* Move override_csp_to_allow_acuant spec to acuant_concern_spec.rb

* Stopped using doc_auth_enable_presigned_s3_urls flag

Async document capture has not been used.

* Remove hybrid_flow_spec

Replaced by hybrid_mobile/hybrid_mobile_spec.rb

* Remove more async specs

* Changelog

changelog: internal,dismantle flow state machine,remove old hybrid flow

* Remove idv_doc_auth_verify_document_status_submitted method, unused

* Removed ApiDocumentverificationStatusForm

* Put ApiDocumentverificationStatusForm back

To be removed in LG-9611

---------
