--- app/services/idv/flows/capture_doc_flow.rb
@@ -1,37 +1,0 @@
-module Idv
-  module Flows
-    class CaptureDocFlow < Flow::BaseFlow
-      STEPS = {
-        document_capture: Idv::Steps::DocumentCaptureStep,
-        capture_complete: Idv::Steps::CaptureCompleteStep,
-      }.freeze
-
-      STEP_INDICATOR_STEPS = [
-        { name: :getting_started },
-        { name: :verify_id },
-        { name: :verify_info },
-        { name: :verify_phone_or_address },
-        { name: :secure_account },
-      ].freeze
-
-      ACTIONS = {
-        verify_document_status: Idv::Actions::VerifyDocumentStatusAction,
-      }.freeze
-
-      def initialize(controller, session, _name)
-        super(controller, STEPS, ACTIONS, session)
-      end
-
-      def flow_path
-        'hybrid'
-      end
-
-      def extra_analytics_properties
-        {
-          acuant_sdk_upgrade_ab_test_bucket:
-            AbTests::ACUANT_SDK.bucket(flow_session[:document_capture_session_uuid]),
-        }
-      end
-    end
-  end
-end
