+++ app/services/idv/flows/doc_auth_flow.rb
@@ -29,7 +29,6 @@ module Idv
       ACTIONS = {
         cancel_link_sent: Idv::Actions::CancelLinkSentAction,
         redo_document_capture: Idv::Actions::RedoDocumentCaptureAction,
-        verify_document_status: Idv::Actions::VerifyDocumentStatusAction,
       }.freeze
 
       attr_reader :idv_session # this is needed to support (and satisfy) the current LOA3 flow
