--- app/views/idv/capture_doc/document_capture.html.erb
@@ -1,14 +1,0 @@
-<%= render(
-      'idv/shared/document_capture',
-      document_capture_session_uuid: document_capture_session_uuid,
-      flow_path: 'hybrid',
-      sp_name: decorated_session.sp_name,
-      failure_to_proof_url: idv_capture_doc_return_to_sp_url,
-      front_image_upload_url: front_image_upload_url,
-      back_image_upload_url: back_image_upload_url,
-      acuant_sdk_upgrade_a_b_testing_enabled: acuant_sdk_upgrade_a_b_testing_enabled,
-      use_alternate_sdk: use_alternate_sdk,
-      acuant_version: acuant_version,
-      in_person_cta_variant_testing_enabled: in_person_cta_variant_testing_enabled,
-      in_person_cta_variant_active: in_person_cta_variant_active,
-    ) %>
