+++ app/views/idv/hybrid_mobile/capture_complete/show.html.erb
@@ -1,6 +1,6 @@
 <% content_for(:pre_flash_content) do %>
   <%= render StepIndicatorComponent.new(
-        steps: Idv::Flows::CaptureDocFlow::STEP_INDICATOR_STEPS,
+        steps: Idv::Flows::DocAuthFlow::STEP_INDICATOR_STEPS,
         current_step: :verify_id,
         locale_scope: 'idv',
         class: 'margin-x-neg-2 margin-top-neg-4 tablet:margin-x-neg-6 tablet:margin-top-neg-4',
