+++ spec/controllers/idv/doc_auth_controller_spec.rb
@@ -209,210 +209,6 @@ describe Idv::DocAuthController do
     end
   end
 
-  describe 'async document verify status' do
-    before do
-      mock_document_capture_step
-    end
-    let(:good_pii) do
-      {
-        first_name: Faker::Name.first_name,
-        last_name: Faker::Name.last_name,
-        dob: (Time.zone.today - (IdentityConfig.store.idv_min_age_years + 1).years).to_s,
-        address1: Faker::Address.street_address,
-        city: Faker::Address.city,
-        state: Faker::Address.state_abbr,
-        zipcode: Faker::Address.zip_code,
-        state_id_type: 'drivers_license',
-        state_id_number: '111',
-        state_id_jurisdiction: 'WI',
-      }
-    end
-    let(:good_result) do
-      {
-        success: true,
-        errors: {},
-        messages: ['message'],
-        pii_from_doc: good_pii,
-        attention_with_barcode: false,
-      }
-    end
-    let(:bad_pii) do
-      {
-        first_name: Faker::Name.first_name,
-        last_name: nil,
-        dob: nil,
-        address1: Faker::Address.street_address,
-        city: Faker::Address.city,
-        state: Faker::Address.state_abbr,
-        zipcode: Faker::Address.zip_code,
-        state_id_type: 'drivers_license',
-        state_id_number: '111',
-        state_id_jurisdiction: 'WI',
-      }
-    end
-    let(:bad_pii_result) do
-      {
-        success: true,
-        errors: {},
-        messages: ['message'],
-        pii_from_doc: bad_pii,
-        attention_with_barcode: false,
-      }
-    end
-    let(:fail_result) do
-      {
-        pii_from_doc: {},
-        success: false,
-        errors: { front: 'Wrong document' },
-        messages: ['message'],
-        attention_with_barcode: false,
-      }
-    end
-    let(:hard_fail_result) do
-      {
-        pii_from_doc: {},
-        success: false,
-        errors: { front: 'Wrong document' },
-        messages: ['message'],
-        attention_with_barcode: false,
-        doc_auth_result: 'Failed',
-      }
-    end
-
-    it 'returns status of success' do
-      set_up_document_capture_result(
-        uuid: document_capture_session_uuid,
-        idv_result: good_result,
-      )
-      put :update,
-          params: {
-            step: 'verify_document_status',
-            document_capture_session_uuid: document_capture_session_uuid,
-          }
-
-      expect(response.status).to eq(200)
-      expect(response.body).to eq({ success: true }.to_json)
-    end
-
-    it 'returns status of in progress' do
-      set_up_document_capture_result(
-        uuid: document_capture_session_uuid,
-        idv_result: nil,
-      )
-      put :update,
-          params: {
-            step: 'verify_document_status',
-            document_capture_session_uuid: document_capture_session_uuid,
-          }
-
-      expect(response.status).to eq(202)
-      expect(response.body).to eq({ success: true }.to_json)
-    end
-
-    it 'returns status of fail' do
-      set_up_document_capture_result(
-        uuid: document_capture_session_uuid,
-        idv_result: fail_result,
-      )
-      put :update,
-          params: {
-            step: 'verify_document_status',
-            document_capture_session_uuid: document_capture_session_uuid,
-          }
-
-      expect(response.status).to eq(400)
-      expect(response.body).to eq(
-        {
-          success: false,
-          errors: [{ field: 'front', message: 'Wrong document' }],
-          remaining_attempts: IdentityConfig.store.doc_auth_max_attempts,
-          ocr_pii: nil,
-          result_failed: false,
-        }.to_json,
-      )
-    end
-
-    it 'returns status of hard fail' do
-      set_up_document_capture_result(
-        uuid: document_capture_session_uuid,
-        idv_result: hard_fail_result,
-      )
-      put :update,
-          params: {
-            step: 'verify_document_status',
-            document_capture_session_uuid: document_capture_session_uuid,
-          }
-
-      expect(response.status).to eq(400)
-      expect(response.body).to eq(
-        {
-          success: false,
-          errors: [{ field: 'front', message: 'Wrong document' }],
-          remaining_attempts: IdentityConfig.store.doc_auth_max_attempts,
-          ocr_pii: nil,
-          result_failed: true,
-        }.to_json,
-      )
-    end
-
-    it 'returns status of fail with incomplete PII from doc auth' do
-      set_up_document_capture_result(
-        uuid: document_capture_session_uuid,
-        idv_result: bad_pii_result,
-      )
-
-      expect(@analytics).to receive(:track_event).with(
-        'IdV: doc auth image upload vendor pii validation', include(
-          errors: include(
-            pii: [I18n.t('doc_auth.errors.general.no_liveness')],
-          ),
-          error_details: { pii: [I18n.t('doc_auth.errors.general.no_liveness')] },
-          attention_with_barcode: false,
-          success: false,
-          remaining_attempts: IdentityConfig.store.doc_auth_max_attempts,
-          flow_path: 'standard',
-          pii_like_keypaths: [[:pii]],
-          user_id: nil,
-        )
-      )
-
-      expect(@analytics).to receive(:track_event).with(
-        'IdV: doc auth verify_document_status submitted', include(
-          errors: include(
-            pii: [I18n.t('doc_auth.errors.general.no_liveness')],
-          ),
-          error_details: { pii: [I18n.t('doc_auth.errors.general.no_liveness')] },
-          attention_with_barcode: false,
-          success: false,
-          remaining_attempts: IdentityConfig.store.doc_auth_max_attempts,
-          step: 'verify_document_status',
-          flow_path: 'standard',
-          step_count: 1,
-          pii_like_keypaths: [[:pii]],
-          doc_auth_result: nil,
-        )
-      )
-
-      put :update,
-          params: {
-            step: 'verify_document_status',
-            document_capture_session_uuid: document_capture_session_uuid,
-          }
-
-      expect(response.status).to eq(400)
-      expect(response.body).to eq(
-        {
-          success: false,
-          errors: [{ field: 'pii',
-                     message: I18n.t('doc_auth.errors.general.no_liveness') }],
-          remaining_attempts: IdentityConfig.store.doc_auth_max_attempts,
-          ocr_pii: nil,
-          result_failed: false,
-        }.to_json,
-      )
-    end
-  end
-
   def mock_next_step(step)
     allow_any_instance_of(Idv::Flows::DocAuthFlow).to receive(:next_step).and_return(step)
   end
