+++ spec/controllers/idv/hybrid_mobile/capture_complete_controller_spec.rb
@@ -25,6 +25,8 @@ describe Idv::HybridMobile::CaptureCompleteController do
   before do
     session[:doc_capture_user_id] = user&.id
     session[:document_capture_session_uuid] = document_capture_session_uuid
+    stub_analytics
+    allow(@analytics).to receive(:track_event)
     allow(subject).to receive(:confirm_document_capture_session_complete).
       and_return(true)
   end
@@ -45,45 +47,36 @@ describe Idv::HybridMobile::CaptureCompleteController do
     end
   end
 
-  context 'when doc_auth_hybrid_mobile_controllers_enabled' do
-    before do
-      allow(IdentityConfig.store).to receive(:doc_auth_hybrid_mobile_controllers_enabled).
-        and_return(true)
-      stub_analytics
-      allow(@analytics).to receive(:track_event)
+  describe '#show' do
+    let(:analytics_name) { 'IdV: doc auth capture_complete visited' }
+    let(:analytics_args) do
+      {
+        acuant_sdk_upgrade_ab_test_bucket: :default,
+        analytics_id: 'Doc Auth',
+        flow_path: 'hybrid',
+        irs_reproofing: false,
+        step: 'capture_complete',
+      }
     end
 
-    describe '#show' do
-      let(:analytics_name) { 'IdV: doc auth capture_complete visited' }
-      let(:analytics_args) do
-        {
-          acuant_sdk_upgrade_ab_test_bucket: :default,
-          analytics_id: 'Doc Auth',
-          flow_path: 'hybrid',
-          irs_reproofing: false,
-          step: 'capture_complete',
-        }
-      end
+    it 'renders the show template' do
+      get :show
 
-      it 'renders the show template' do
-        get :show
+      expect(response).to render_template :show
+    end
 
-        expect(response).to render_template :show
-      end
+    it 'sends analytics_visited event' do
+      get :show
 
-      it 'sends analytics_visited event' do
-        get :show
+      expect(@analytics).to have_received(:track_event).with(analytics_name, analytics_args)
+    end
 
-        expect(@analytics).to have_received(:track_event).with(analytics_name, analytics_args)
-      end
+    it 'updates DocAuthLog capture_complete_view_count' do
+      doc_auth_log = DocAuthLog.create(user_id: user.id)
 
-      it 'updates DocAuthLog capture_complete_view_count' do
-        doc_auth_log = DocAuthLog.create(user_id: user.id)
-
-        expect { get :show }.to(
-          change { doc_auth_log.reload.capture_complete_view_count }.from(0).to(1),
-        )
-      end
+      expect { get :show }.to(
+        change { doc_auth_log.reload.capture_complete_view_count }.from(0).to(1),
+      )
     end
   end
 end
