+++ spec/features/idv/proofing_components_spec.rb
@@ -26,17 +26,6 @@ RSpec.describe 'proofing components' do
       acknowledge_and_confirm_personal_key
     end
 
-    context 'async proofing', js: true do
-      let(:ruby_workers_idv_enabled) { true }
-      let(:doc_auth_enable_presigned_s3_urls) { true }
-
-      it 'records proofing components' do
-        proofing_components = user.active_profile.proofing_components
-        expect(proofing_components['document_check']).to eq('mock')
-        expect(proofing_components['document_type']).to eq('state_id')
-      end
-    end
-
     context 'sync proofing', js: true do
       let(:ruby_workers_idv_enabled) { false }
       let(:doc_auth_enable_presigned_s3_urls) { false }
